requirements.txt
azure-functions
azure-storage-blob
aenum==2.2.6
aiofiles==0.6.0
appdirs==1.4.4
certifi==2020.12.5
click==7.1.2
distlib==0.3.1
fastapi==0.63.0
gremlinpython==3.4.6
h11==0.12.0
uvicorn==0.13.4
gunicorn==20.0.4
isodate==0.6.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pipenv==2020.11.15
pydantic==1.8.1
pyparsing==2.4.7
six==1.15.0
starlette==0.13.6
tornado==4.4.1
typing-extensions==3.7.4.3